Actually got all this stuff building again
git-svn-id: http://svn.drobilla.net/lad/grauph@21 a436a847-0d15-0410-975c-d299462d15a1

+++ src/progs/demolition/DemolitionClientInterface.h
@@ -43,7 +43,7 @@ public:
 	void bundle_begin() {}
 	void bundle_end()   {}
 
-	void num_plugins(size_t num) {}
+	void num_plugins(uint32_t num) {}
 
 	// OSC thread functions
 	void error(const string& msg);
